backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.40.0
google-adk==1.0.0
motor==3.6.0
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.9.2
mcp==1.1.2
httpx==0.27.2
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.12
